fix: restore private access control; inline poll loop back into RunnerStore.swift
